The timing tests are complete as well as some gifs to show the results.  Still some minor problems in Euler.

+++ SharedSwept/plotSolutions.py
@@ -29,10 +29,10 @@ import numpy as np
 import subprocess as sp
 import shlex
 import os
+import sys
 import Tkinter as Tk
 import pandas as pd
-from RegressionTesting.fig2Dat import fig2img
-import images2gif as img
+import palettable as pal
 
 alpha = 8.418e-5
 dx = 0.001
@@ -49,6 +49,8 @@ OPT_PREC = [
     "Single",
     "Double"
 ]
+
+#def heat_msg()
 
 #It's kinda getting there.
 master = Tk.Tk()
@@ -193,7 +195,7 @@ if not op.isdir(binpath):
 if not op.isdir(basepath):
     os.mkdir(basepath)
 
-Varfile = op.join(basepath, typename + "_1D_Result.dat")
+Varfile = op.join(basepath, typename + "_Result.dat")
 gitpath = op.dirname(sourcepath)
 gifpath = op.join(op.join(gitpath,'ResultPlots'),'Gifs')
 giffile = op.join(gifpath,typename+".gif")
@@ -203,8 +205,6 @@ if not op.isdir(temppath):
     os.mkdir(temppath)
 
 avifile = op.join(temppath,typename+".avi")
-
-print gifpath
 
 div = 2**divpow.get()
 bks = 2**blkpow.get()
@@ -257,20 +257,33 @@ fin = tuple(f)
 ar = [float(n) for n in fin[0].split()]
 xax = np.linspace(0,ar[0],ar[1])
 ed = ar[0]
-mx = ar[1]/2048.0 * 50.0
+
 dMain = []
 
 for p in range(1,len(fin)):
     ar = fin[p].split()
     dMain.append([ar[0]] + [float(n) for n in ar[1:]])
 
+mx = max(dMain[0][1:])
+mxx = mx/10.
+if mxx < 5:
+    mxx = 5
+
+mxxx = mx + mxx
+f.close()
+
+
 simF = pd.DataFrame(dMain)
 simF = simF.set_index(0)
 typ = simF.index.get_level_values(0).unique()
-print typ
 cross = simF.xs( typ[0] )
 
 cnt = len(cross.index)
+
+if "city" in typ[0]:
+    lw = 2
+else:
+    lw = 4
 
 if cnt < 6:
 
@@ -278,7 +291,7 @@ if cnt < 6:
 
     if len(typ) < 2:
 
-        fig, ax = plt.subplots(figsize=(14.,8.))
+        fig, ax = plt.subplots(figsize=(8.,6.))
 
         df_sim = cross.set_index(1)
         df_sim.columns = xax
@@ -291,18 +304,19 @@ if cnt < 6:
         ax.grid(alpha=0.5)
         ax.set_xlabel("Spatial point")
         ax.set_ylabel(typ[0])
-        ax.set_ylim([0,mx])
+        ax.set_ylim([-5,mxxx])
         ax.set_xlim([0,ed])
 
         for tfs in cl:
-            ax.plot(df_sim['index'], df_sim[tfs], label="{:.2f} (s)".format(tfs))
+            ax.plot(df_sim['index'], df_sim[tfs], label="{:.2f} (s)".format(tfs), linewidth=lw)
 
+        plt.tight_layout()
         ax.legend()
         plt.show()
 
     else:
 
-        fig, ax = plt.subplots((2,2),figsize=(14.,8.))
+        fig, ax = plt.subplots(2, 2 ,figsize=(14.,8.))
         ax = ax.ravel()
         plt.suptitle(typename + ' | {0} spatial points   '.format(int(div)), fontsize="medium")
 
@@ -321,16 +335,16 @@ if cnt < 6:
             ax[i].set_title(ty)
 
             for tfs in cl:
-                ax[i].plot(df_sim['index'], df_sim[tfs], label="{:.2f} (s)".format(tfs))
+                ax[i].plot(df_sim['index'], df_sim[tfs], label="{:.2f} (s)".format(tfs), linewidth=2)
 
         hand, lbl = ax[0].get_legend_handles_labels()
         fig.legend(hand, lbl, 'upper_right', fontsize="medium")
         plt.tight_layout(pad=0.2, w_pad=0.75, h_pad=1.5)
         plt.subplots_adjust(bottom=0.08, right=0.82, top=0.92)
-        plt.suptitle(typename + ' | {0} spatial points   '.format(int(div)), fontsize="medium")
         plt.show()
 
 else:
+    print "Making GIF"
     os.chdir(temppath)
     #If it's not euler
     if len(typ) == 1:
@@ -341,7 +355,8 @@ else:
         cl = df_sim.columns.values.tolist()
         df_sim.reset_index(inplace=True)
 
-        fig, ax = plt.subplots(figsize=(14.,8.))
+
+        fig, ax = plt.subplots(figsize=(8.,6.))
         plt.hold(False)
         fig.suptitle(typename + ' | {0} spatial points   '.format(int(div)), fontweight='bold')
 
@@ -349,54 +364,67 @@ else:
 
         for k,tfs in enumerate(cl):
 
-            ax.plot(df_sim['index'], df_sim[tfs])
+            ax.plot(df_sim['index'], df_sim[tfs], linewidth=lw)
 
             ax.set_title("{:.2f} (s)".format(tfs))
             ax.set_xlabel("Spatial point")
             ax.grid(alpha=0.5)
             ax.set_ylabel(typ[0])
-            ax.set_ylim([0,mx])
+            ax.set_ylim([-5,mxxx])
             ax.set_xlim([0,ed])
 
             plt.savefig("frame"+str(k))
 
-
     else:
-        simF.reset_index(inplace=True)
-        df_sim = simF.set_index(1)
-        tfidx = simF.index.get_level_values(0).unique()
 
-        fig, ax = plt.subplots((2,2),figsize=(14.,8.))
+        simF.reset_index(inplace=True)
+        tfidx = simF[1].unique()
+        df_sim = simF.set_index(1)
+
+        fig, ax = plt.subplots(2, 2 ,figsize=(14.,8.))
         ax = ax.ravel()
-        plt.suptitle(typename + ' | {0} spatial points   '.format(int(div)), fontweight='bold')
-        fig.subplots_adjust(top=0.85)
+        fig.suptitle(typename + ' | {0} spatial points   '.format(int(div)), fontweight='bold')
 
         for k,tf in enumerate(tfidx):
+            df_simz = df_sim.xs( tf )
+            df_simz = df_simz.set_index(0)
+            df_simz.columns = xax
+            df_simz = df_simz.transpose()
 
-            df_sim = simF.xs( tf )
-            df_sim = df_sim.set_index(0)
-            df_sim.columns = xax
-            df_sim = df_sim.transpose()
-            cl = df_sim.columns.values.tolist()
-            df_sim.reset_index(inplace=True)
-            fig.title('{:.2e} s '.format(tf))
+            cl = df_simz.columns.values.tolist()
+            df_simz.reset_index(inplace=True)
+
+            txt = fig.text(.85, .85, "{:.2e} s ".format(tf), bbox={'facecolor':'white'})
 
             for i, state in enumerate(cl):
-                ax[i].plot(df_sim['index'], df_sim[state])
+                ax[i].plot(df_simz['index'], df_simz[state], linewidth=2)
+                ax[i].hold(False)
                 ax[i].grid(alpha=0.5)
                 ax[i].set_xlabel("Spatial point")
-                ax[i].set_title(typ[n])
-                if typ[n] == 'energy':
+                ax[i].set_title(state)
+                if 'ergy' in state:
                     ax[i].set_ylim([1.5,3.5])
                 else:
                     ax[i].set_ylim([0.0,1.1])
+                plt.tight_layout(pad=0.2, w_pad=0.75, h_pad=1.5)
+                plt.subplots_adjust(bottom=0.08, right=0.82, top=0.92)
 
             plt.savefig("frame"+str(k))
+            txt.remove()
 
+    st = 'linux'
+    if st in sys.platform:
+        try:
+            sp.call(['ffmpeg', '-i', 'frame%d.png', '-r', '4', avifile])
+            sp.call(['ffmpeg', '-i', avifile, '-t', '5', giffile])
+        except:
+            print '------------------'
+            print 'Install ffmpeg with: sudo apt-get install ffmpeg'
+            raise SystemExit
 
-    sp.call(['ffmpeg', '-i', 'frame%d.png', '-r', '4', avifile])
-    sp.call(['ffmpeg', '-i', 'output.avi', '-t', '5', giffile])
-
-    f = os.listdir(".")
-    for fm in f:
-        os.remove(fm)
+        f = os.listdir(".")
+        for fm in f:
+            os.remove(fm)
+    else:
+        print '------------------'
+        print 'This script only makes gifs on linux with ffmpeg.  The images are still in the folder under ResultPlots/Gifs/Temp.'
